Added basic damage percentage Minor improvements to Match HUD Minor improvements to Junkbot animations

+++ PDM Charlie/Assets/Scripts/HUD/PlayerMatchInfoController.cs
@@ -14,7 +14,7 @@ public class PlayerMatchInfoController : MonoBehaviour
         this.playerName.text = newName;   
     }
 
-    public void UpdatePlayerPercentage(int newPercentage)
+    public void UpdatePlayerPercentage(float newPercentage)
     {
         this.playerPercentage.text = $"{newPercentage}%";
     }
